+++ modules/mod_quicklook_v100.py
@@ -16,7 +16,7 @@ from common.functions import cocoa_time
 from common.functions import read_stream_bplist
 from common.functions import query_db
 
-# IMPORT STATIC VARIABLES FROM MACXTR
+# IMPORT STATIC VARIABLES FROM MAIN
 from __main__ import inputdir
 from __main__ import outputdir
 from __main__ import forensic_mode