+++ modules/mod_spotlight_v100.py
@@ -15,7 +15,7 @@ well as the timestamp that it was last opened.
 # IMPORT FUNCTIONS FROM COMMON.FUNCTIONS
 from common.functions import stats2
 
-# IMPORT STATIC VARIABLES FROM MACXTR
+# IMPORT STATIC VARIABLES FROM MAIN
 from __main__ import inputdir
 from __main__ import outputdir
 from __main__ import forensic_mode